Move PAKE logic to separate module
