+++ tests/ClientProtocol.hs
@@ -3,24 +3,22 @@ module ClientProtocol (tests) where
 
 import Protolude hiding (phase)
 
-import Hedgehog (forAll, property, tripping, (===))
+import Hedgehog (forAll, property, (===))
 import qualified Hedgehog.Gen as Gen
 import qualified Hedgehog.Range as Range
 import Test.Tasty (TestTree, testGroup)
 import Test.Tasty.Hedgehog (testProperty)
 
 import qualified MagicWormhole.Internal.ClientProtocol as ClientProtocol
+import qualified MagicWormhole.Internal.Pake as Pake
 
 
 tests :: IO TestTree
 tests = pure $ testGroup "Peer"
-  [ testProperty "SPAKE2 messages roundtrip" $ property $ do
-      element <- forAll $ Gen.bytes (Range.singleton 32)
-      tripping element ClientProtocol.spakeBytesToMessageBody ClientProtocol.messageBodyToSpakeBytes
-  , testProperty "SecretBox encryption roundtrips" $ property $ do
+  [ testProperty "SecretBox encryption roundtrips" $ property $ do
       purpose <- forAll $ Gen.bytes (Range.linear 0 10)
       secret <- forAll $ Gen.bytes (Range.linear 0 10)
-      let key = ClientProtocol.deriveKey (ClientProtocol.SessionKey secret) purpose
+      let key = ClientProtocol.deriveKey (Pake.SessionKey secret) purpose
       plaintext <- forAll $ Gen.bytes (Range.linear 1 256)
       ciphertext <- liftIO $ ClientProtocol.encrypt key plaintext
       let decrypted = ClientProtocol.decrypt key ciphertext
